services/researcher/requirements.txt
langchain
langchain-community
langchain-google-genai
tavily-python==0.3.0
kafka-python
pydantic
pydantic-settings
structlog
python-json-logger
openai
groq
google-generativeai
prometheus-client==0.19.0
psycopg2-binary
sqlalchemy
redis
# Multi-provider research dependencies
duckduckgo-search==5.0.0
requests==2.31.0
loguru==0.7.2
